#101 rename: 테스트 코드 이전
