--- src/test/java/codesquad/fineants/spring/api/portfolio_stock/PortfolioHoldingServiceTest.java
@@ -1,205 +1,0 @@
-package codesquad.fineants.spring.api.portfolio_stock;
-
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.ActiveProfiles;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import codesquad.fineants.domain.member.Member;
-import codesquad.fineants.domain.member.MemberRepository;
-import codesquad.fineants.domain.oauth.support.AuthMember;
-import codesquad.fineants.domain.portfolio.Portfolio;
-import codesquad.fineants.domain.portfolio.PortfolioRepository;
-import codesquad.fineants.domain.portfolio_gain_history.PortfolioGainHistoryRepository;
-import codesquad.fineants.domain.portfolio_holding.PortfolioHolding;
-import codesquad.fineants.domain.portfolio_holding.PortfolioHoldingRepository;
-import codesquad.fineants.domain.purchase_history.PurchaseHistory;
-import codesquad.fineants.domain.purchase_history.PurchaseHistoryRepository;
-import codesquad.fineants.domain.stock.Market;
-import codesquad.fineants.domain.stock.Stock;
-import codesquad.fineants.domain.stock.StockRepository;
-import codesquad.fineants.spring.api.errors.exception.NotFoundResourceException;
-import codesquad.fineants.spring.api.portfolio_stock.event.publisher.PortfolioHoldingEventPublisher;
-import codesquad.fineants.spring.api.portfolio_stock.request.PortfolioStockCreateRequest;
-import codesquad.fineants.spring.api.portfolio_stock.response.PortfolioStockCreateResponse;
-import codesquad.fineants.spring.api.portfolio_stock.response.PortfolioStockDeleteResponse;
-import codesquad.fineants.spring.api.portfolio_stock.service.PortfolioStockService;
-
-@ActiveProfiles("test")
-@SpringBootTest
-class PortfolioHoldingServiceTest {
-	@Autowired
-	private PortfolioStockService portfolioStockService;
-
-	@Autowired
-	private PortfolioHoldingRepository portFolioHoldingRepository;
-
-	@Autowired
-	private PortfolioRepository portfolioRepository;
-
-	@Autowired
-	private MemberRepository memberRepository;
-
-	@Autowired
-	private StockRepository stockRepository;
-
-	@Autowired
-	private ObjectMapper objectMapper;
-
-	@Autowired
-	private PurchaseHistoryRepository purchaseHistoryRepository;
-
-	@Autowired
-	private PortfolioGainHistoryRepository portfolioGainHistoryRepository;
-
-	@MockBean
-	private PortfolioHoldingEventPublisher publisher;
-
-	private Member member;
-
-	private Portfolio portfolio;
-
-	private Stock stock;
-
-	@BeforeEach
-	void init() {
-		Member member = Member.builder()
-			.nickname("일개미1234")
-			.email("kim1234@gmail.com")
-			.password("kim1234@")
-			.provider("local")
-			.build();
-		this.member = memberRepository.save(member);
-
-		Portfolio portfolio = Portfolio.builder()
-			.name("내꿈은 워렌버핏")
-			.securitiesFirm("토스")
-			.budget(1000000L)
-			.targetGain(1500000L)
-			.maximumLoss(900000L)
-			.member(member)
-			.build();
-		this.portfolio = portfolioRepository.save(portfolio);
-
-		Stock stock = Stock.builder()
-			.companyName("삼성전자보통주")
-			.tickerSymbol("005930")
-			.companyNameEng("SamsungElectronics")
-			.stockCode("KR7005930003")
-			.market(Market.KOSPI)
-			.build();
-		this.stock = stockRepository.save(stock);
-	}
-
-	@AfterEach
-	void tearDown() {
-		purchaseHistoryRepository.deleteAllInBatch();
-		portFolioHoldingRepository.deleteAllInBatch();
-		portfolioGainHistoryRepository.deleteAllInBatch();
-		portfolioRepository.deleteAllInBatch();
-		memberRepository.deleteAllInBatch();
-		stockRepository.deleteAllInBatch();
-	}
-
-	@DisplayName("사용자는 포트폴리오에 종목을 추가한다")
-	@Test
-	void addPortfolioStock() throws JsonProcessingException {
-		// given
-		Long portfolioId = portfolio.getId();
-		Map<String, Object> requestBodyMap = new HashMap<>();
-		requestBodyMap.put("tickerSymbol", stock.getTickerSymbol());
-		PortfolioStockCreateRequest request = objectMapper.readValue(
-			objectMapper.writeValueAsString(requestBodyMap), PortfolioStockCreateRequest.class);
-
-		// when
-		PortfolioStockCreateResponse response = portfolioStockService.addPortfolioStock(portfolioId,
-			request, AuthMember.from(member));
-
-		// then
-		assertAll(
-			() -> assertThat(response)
-				.extracting("portfolioStockId")
-				.isNotNull(),
-			() -> assertThat(portFolioHoldingRepository.findAll()).hasSize(1)
-		);
-	}
-
-	@DisplayName("사용자는 포트폴리오에 존재하지 않는 종목을 추가할 수 없다")
-	@Test
-	void addPortfolioStockWithNotExistStock() throws JsonProcessingException {
-		// given
-		Long portfolioId = portfolio.getId();
-		Map<String, Object> requestBodyMap = new HashMap<>();
-		requestBodyMap.put("tickerSymbol", "999999");
-		PortfolioStockCreateRequest request = objectMapper.readValue(
-			objectMapper.writeValueAsString(requestBodyMap), PortfolioStockCreateRequest.class);
-
-		// when
-		Throwable throwable = catchThrowable(() -> portfolioStockService.addPortfolioStock(portfolioId,
-			request, AuthMember.from(member)));
-
-		// then
-		assertThat(throwable).isInstanceOf(NotFoundResourceException.class)
-			.extracting("message")
-			.isEqualTo("종목을 찾을 수 없습니다");
-	}
-
-	@DisplayName("사용자는 포트폴리오의 종목을 삭제한다")
-	@Test
-	void deletePortfolioStock() {
-		// given
-		PortfolioHolding portfolioHolding = portFolioHoldingRepository.save(
-			PortfolioHolding.empty(portfolio, stock)
-		);
-
-		purchaseHistoryRepository.save(
-			PurchaseHistory.builder()
-				.purchaseDate(LocalDateTime.now())
-				.purchasePricePerShare(10000.0)
-				.numShares(1L)
-				.memo("첫구매")
-				.build()
-		);
-
-		Long portfolioHoldingId = portfolioHolding.getId();
-		// when
-		PortfolioStockDeleteResponse response = portfolioStockService.deletePortfolioStock(
-			portfolioHoldingId, portfolio.getId(), AuthMember.from(member));
-
-		// then
-		assertAll(
-			() -> assertThat(response).extracting("portfolioHoldingId").isNotNull(),
-			() -> assertThat(portFolioHoldingRepository.findById(portfolioHoldingId).isEmpty()).isTrue(),
-			() -> assertThat(purchaseHistoryRepository.findAllByPortfolioHoldingId(portfolioHoldingId)).isEmpty()
-		);
-	}
-
-	@DisplayName("사용자는 존재하지 않은 포트폴리오의 종목을 삭제할 수 없다")
-	@Test
-	void deletePortfolioStockWithNotExistPortfolioStockId() {
-		// given
-		Long portfolioStockId = 9999L;
-
-		// when
-		Throwable throwable = catchThrowable(() -> portfolioStockService.deletePortfolioStock(
-			portfolioStockId, portfolio.getId(), AuthMember.from(member)));
-
-		// then
-		assertThat(throwable).isInstanceOf(NotFoundResourceException.class).extracting("message")
-			.isEqualTo("포트폴리오 종목이 존재하지 않습니다");
-	}
-}
